% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fast.iasva.R, R/iasva.unit.R
\name{calc.rsq}
\alias{calc.rsq}
\alias{calc.rsq}
\title{calc.rsq}
\usage{
calc.rsq(resid, fit)

calc.rsq(resid, fit)
}
\arguments{
\item{resid}{residual matrix}

\item{fit}{output of .lm.fit()}

\item{resid}{residual matrix}

\item{fit}{output of .lm.fit()}
}
\value{
R squared

R squared
}
\description{
calc.rsq

calc.rsq
}
