\name{counts}
\docType{data}
\alias{counts}
\title{Group Category}
\description{
    This subset of a single-cell gene expression dataset gathered
    from the islets of diabetic and non-diabetic patients measures
    the gene expresssion of 14416 genes from 101 cells.
}
\usage{counts}
\format{
    A dataframe containing 101 observations and 14416 features.
}
