% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fast.iasva.unit.R
\name{fast.iasva.unit}
\alias{fast.iasva.unit}
\title{A function for estimating a surrogate variable (fast IA-SVA)}
\usage{
fast.iasva.unit(Y, X, intercept = TRUE, component.retention = TRUE,
  num.sv.retention = NULL, tol = 1e-10, verbose = FALSE)
}
\arguments{
\item{Y}{read counts matrix with samples in row and genes in column.}

\item{X}{known variables.}

\item{intercept}{If intercept=FALSE, the linear intercept is not included in the model.}

\item{component.retention}{If component.retention=TRUE, the percentage of unmodeled variance explained by the putative hidden factor.}

\item{num.sv.retention}{num of top singular values to be used in computing the percentage of unmodeled variation explained by the putative hidden factor. If num.sv.retention=NULL, all singular values are used.}

\item{tol}{stopping tolerance for the augmented implicitly restarted Lanczos bidiagonalization algorithm}

\item{verbose}{If verbose=TRUE, the function outputs detailed messages.}
}
\value{
sv estimated surrogate variable.

percentage percentage of unmodeled variance explained by the estimated surrogate variable.
}
\description{
fast.iasva.unit() estimates a hidden factor for unwanted variation while accounting for all known factors and computes 
its contribution (i.e., the percentage of unmodeled variation explained by the hidden factor) on the unmodeled variation in the data. 
This function is called in fast.iasva() function to iteratively identify hidden factors.
}
