% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find.markers.R
\name{find.markers}
\alias{find.markers}
\title{A function for finding markers for hidden factors}
\usage{
find.markers(Y, iasva.sv, method = "BH", sig.cutoff = 0.05,
  rsq.cutoff = 0.3, verbose = FALSE)
}
\arguments{
\item{Y}{read counts matrix with samples in row and genes in column.}

\item{iasva.sv}{matrix of estimated surrogate variables, one column for each surrogate variable.}

\item{method}{multiple testing adjustment method, default="BH".}

\item{sig.cutoff}{significance cutoff.}

\item{rsq.cutoff}{R squared cutoff.}

\item{verbose}{If verbose=TRUE, the function outputs detailed messages.}
}
\value{
marker.counts read counts matrix of markers, one column for each cell.
}
\description{
find.markers() takes a read counts matrix of entire gene set and a matrix of surrogate variables estimated by IA-SVA as input, 
identifies marker genes highly correlated with each surrogate variable and returns a read counts matrix of the markers.
}
\examples{
library(iasva)
data(iasva_test)
iasva.res <- iasva(iasvaY, iasvaX) 
markers <- find.markers(iasvaY, iasva.res$sv)

}
