% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iasva.unit.R
\name{iasva.unit}
\alias{iasva.unit}
\title{A function for estimating a surrogate variable}
\usage{
iasva.unit(Y, X, permute = TRUE, num.p = 100, intercept = TRUE,
  verbose = FALSE)
}
\arguments{
\item{Y}{The read counts matrix with samples in row and genes in column.}

\item{X}{The known variables including the primary variables of interest.}

\item{permute}{If permute=TRUE, a permutation test (Buja and Eyuboglu 1992, Leek and Storey 2008) is conducted to assess the significance of the putative hidden factor.}

\item{num.p}{The number of permutations that will be used to calculate the permuation test p-value.}

\item{intercept}{If intercept=FALSE, the linear intercept is not included in the model.}

\item{verbose}{If verbose=TRUE, the function outputs detailed messages.}
}
\value{
sv The matrix of estimated surrogate variables, one column for each surrogate variable.

sv.resid The matrix of residuals.

pc.stat.obs The vector of PC test statistic values, one value for each surrogate variable.

pval The vector of permuation p-values, one value for each surrogate variable.

wgt The matrix of gene weights (0-1 normalized R-squared values), one column for each surrogate variable.

rsq The matrix of R-squared values, one column for each surrogate variable.
}
\description{
iasva.unit() estimates a hidden factor for unwanted variation while accounting for all known factors and tests the significance 
of its contribution on the unmodeled variation in the data. This function is called in iasva() function to iteratively identify hidden 
factors.
}
\examples{

}
