% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iasva.unit.R
\name{permute.svd.factory}
\alias{permute.svd.factory}
\title{permute.svd.factory}
\usage{
permute.svd.factory(lY, X, num.sv.permtest, tol, verbose)
}
\arguments{
\item{lY}{log-transformed read counts matrix}

\item{X}{matrix of known factors}

\item{num.sv.permtest}{num of top singluar values to be used in computing the permutation test statistic.}

\item{tol}{stopping tolerance for the augmented implicitly restarted Lanczos bidiagonalization algorithm}

\item{verbose}{If verbose=TRUE, the function outputs detailed messages.}
}
\value{
permute.svd()
}
\description{
permute.svd.factory
}
