% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/study.R2.R
\name{study.R2}
\alias{study.R2}
\title{A function to study different values of R2}
\usage{
study.R2(Y, iasva.sv, selected.svs = 2, no.clusters = 2, verbose = FALSE)
}
\arguments{
\item{Y}{read counts matrix with samples in row and genes in column.}

\item{iasva.sv}{matrix of estimated surrogate variables, one column for each surrogate variable.}

\item{selected.svs}{list of SVs that are selected for the analyses. Default is SV2}

\item{no.clusters}{No of clusters to be used in the analyses. Default is 2.}
}
\value{
a summary plot that represents silhoutte index and marker gene counts as a function of R2 and corresponding matrices.
}
\description{
study.R2() studies how different R2 thresholds is changging: 1) number of marker genes; 2) clustering quality (assuming number of clusters is known). 
It generated diagnostic plots that shows how selected genes and clustering quality changes as a function of R2 threshold.
}
