% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find.markers.R
\name{find.markers}
\alias{find.markers}
\title{A function for finding markers for hidden factors}
\usage{
find.markers(Y, iasva.sv, method = "BH", sig.cutoff = 0.05,
  rsq.cutoff = 0.3, verbose = FALSE)
}
\arguments{
\item{Y}{The read counts matrix with samples in row and genes in column.}

\item{iasva.sv}{The matrix of estimated surrogate variables, one column for each surrogate variable.}

\item{method}{The multiple testing adjustment method, default="BH".}

\item{sig.cutoff}{The cutoff for the significance.}

\item{rsq.cutoff}{The cutoff for the R squared.}

\item{verbose}{If verbose=TRUE, the function outputs detailed messages.}
}
\value{
marker.counts The matrix of read counts for markers, one column for each cell.
}
\description{
find.markers() returns a read counts matrix of markers for each hidden factor detected by IA-SVA.
}
\examples{

}
