% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iasva.R
\name{iasva}
\alias{iasva}
\title{A function for iteratively adjusted surrogate variable analysis (IA-SVA)}
\usage{
iasva(Y, X, permute = TRUE, num.p = 100, num.sv = NULL,
  sig.cutoff = 0.05, intercept = TRUE, verbose = FALSE)
}
\arguments{
\item{Y}{The read counts matrix with samples in row and genes in column.}

\item{X}{The known variables including the primary variables of interest.}

\item{permute}{If permute=TRUE, a permutation test (Buja and Eyuboglu 1992, Leek and Storey 2008) is conducted to assess the significance of the putative hidden factor.}

\item{num.p}{The number of permutations that will be used to calculate the permuation test p-value.}

\item{num.sv}{The number of surrogate variables to estimate.}

\item{sig.cutoff}{The significance threshold for the permutation test}

\item{intercept}{If intercept=FALSE, the linear intercept is not included in the model.}

\item{verbose}{If verbose=TRUE, the function outputs detailed messages.}
}
\value{
sv The matrix of estimated surrogate variables, one column for each surrogate variable.

sv.resid The matrix of residuals.

pc.stat.obs The vector of PC test statistic values, one value for each surrogate variable.

pval The vector of permuation p-values, one value for each surrogate variable.

wgt The matrix of gene weights (0-1 normalized R-squared values), one column for each surrogate variable.

rsq The matrix of R-squared values, one column for each surrogate variable.

n.sv The number of significant/obtained surrogate variables.
}
\description{
The iterative procedure of IA-SVA is implemented in this function (iasva). iasva() iteratively runs iasva.unit() function 
to identify a hidden factor for unwanted variation while accounting for all known factors and test the significance 
of its contribution on the unmodeled variation in the data. If the test statistic of detected factor is significant, 
iasva() includes the factor as a known variable in the next iteration to find further hidden factors.
}
\examples{
library(iasva)
data(iasva_test)
iasva(iasvaY, iasvaX)  
}

